+++ .github/workflows/pythontest.yml
@@ -1,0 +1,43 @@
+# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
+# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
+
+name: Python Test
+
+on: 
+  push:
+    branches: 
+      - '**'
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.5, 3.6, 3.7, 3.8]
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v1
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e git+https://github.com/DomAmato/smpp.pdu.git@master#egg=smpp.pdu
+        pip install -r requirements.txt
+        pip install service_identity
+    - name: Lint
+      run: |
+        pip install pylint
+        pylint -rn --errors-only ./smpp 
+    - name: Test
+      env:
+        COVERALLS_REPO_TOKEN: ${{ secrets.COVERALLS_REPO_TOKEN }}
+      run: |
+        pip install coveralls mock
+        coverage run --source=smpp -m twisted.trial tests
+        coveralls
