--- .travis.yml
@@ -1,11 +1,0 @@
-language: python
-
-python:
-  - "2.6"
-  - "2.7"
-
-install:
-  - "pip install -r requirements.txt --use-mirrors"
-  - "pip install -r dev_requirements.txt --use-mirrors"
-
-script: py.test
