+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include README.markdown
+include README.md
