+++ requirements.txt
@@ -1,4 +1,5 @@
-twisted
-enum
-pyOpenSSL
+Twisted~=20.3.0
+pyOpenSSL~=19.1.0
+# cant install from other projects
+# smpp.pdu @ git+https://github.com/DomAmato/smpp.pdu.git@master
 smpp.pdu
