+++ smpp/twisted/client.py
@@ -17,7 +17,7 @@ import logging
 from OpenSSL import SSL
 from twisted.internet.protocol import ClientFactory
 from twisted.internet import defer, reactor, ssl
-from twisted.application import service
+from twisted.application.service import Service
 from smpp.twisted.protocol import SMPPClientProtocol, DataHandlerResponse
 
 LOG_CATEGORY="smpp.twisted.client"
@@ -30,16 +30,17 @@ class SMPPClientFactory(ClientFactory):
         self.config = config
         self.buildProtocolDeferred = defer.Deferred()
         self.log = logging.getLogger(LOG_CATEGORY)
-    
+
     def getConfig(self):
         return self.config
 
     def buildProtocol(self, addr):
         p = ClientFactory.buildProtocol(self, addr)
         #This is a sneaky way of passing the protocol instance back to the caller
+        #pylint: disable=no-member
         reactor.callLater(0, self.buildProtocolDeferred.callback, p)
         return p
-    
+
     def clientConnectionFailed(self, connector, reason):
         """Connection failed
         """
@@ -50,7 +51,7 @@ class CtxFactory(ssl.ClientContextFactory):
 
     def __init__(self, config):
         self.smppConfig = config
-    
+
     def getContext(self):
         self.method = SSL.SSLv23_METHOD
         ctx = ssl.ClientContextFactory.getContext(self)
@@ -58,9 +59,9 @@ class CtxFactory(ssl.ClientContextFactory):
             ctx.use_certificate_file(self.smppConfig.SSLCertificateFile)
         return ctx
 
-class SMPPClientBase(object):
+class SMPPClientBase:
     msgHandler = None
-    
+
     def __init__(self, config):
         self.config = config
         self.log = logging.getLogger(LOG_CATEGORY)
@@ -71,44 +72,46 @@ class SMPPClientBase(object):
         factory = SMPPClientFactory(self.config)
         if self.config.useSSL:
             self.log.warning('Establishing SSL connection to %s:%d' % (self.config.host, self.config.port))
+            #pylint: disable=no-member
             reactor.connectSSL(self.config.host, self.config.port, factory, CtxFactory(self.config))
         else:
             self.log.warning('Establishing TCP connection to %s:%d' % (self.config.host, self.config.port))
-            reactor.connectTCP(self.config.host, self.config.port, factory) 
+            #pylint: disable=no-member
+            reactor.connectTCP(self.config.host, self.config.port, factory)
         return factory.buildProtocolDeferred.addCallback(self.onConnect)
-    
+
     def onConnect(self, smpp):
         self.smpp = smpp
         if self.msgHandler is not None:
             smpp.setDataRequestHandler(self.msgHandler)
         return smpp
-        
+
     def connectAndBind(self):
         self.bindDeferred = defer.Deferred()
         self.connect().addCallback(self.doBind).addErrback(self.bindDeferred.errback)
         return self.bindDeferred
-        
+
     def doBind(self, smpp):
         self.bind(smpp).addCallback(self.bound).addErrback(self.bindFailed, smpp)
         return smpp
-        
+
     def bind(self, smpp):
         raise NotImplementedError()
-        
+
     #If bind fails, don't errback until we're disconnected
     def bindFailed(self, error, smpp):
         smpp.getDisconnectedDeferred().addCallback(lambda result: self.bindDeferred.errback(error))
-        
+
     def bound(self, result):
         self.bindDeferred.callback(result.smpp)
 
 class SMPPClientTransmitter(SMPPClientBase):
-        
+
     def bind(self, smpp):
         return smpp.bindAsTransmitter()
 
 class SMPPClientReceiver(SMPPClientBase):
-    
+
     def __init__(self, config, msgHandler):
         SMPPClientBase.__init__(self, config)
         self.msgHandler = msgHandler
@@ -123,31 +126,31 @@ class SMPPClientTransceiver(SMPPClientReceiver):
 
 #TODO - move this to mozes code base since
 # the service support in Twisted is so crappy
-class SMPPClientService(service.Service):
-    
+class SMPPClientService(Service):
+
     def __init__(self, smppClient):
         self.client = smppClient
         self.stopDeferred = defer.Deferred()
         self.log = logging.getLogger(LOG_CATEGORY)
-        
+
     def getStopDeferred(self):
         return self.stopDeferred
-    
+
     @defer.inlineCallbacks
     def startService(self):
-        service.Service.startService(self)
+        Service.startService(self)
         bindDeferred = self.client.connectAndBind()
         bindDeferred.addErrback(self.handleStartError)
         smpp = yield bindDeferred
         smpp.getDisconnectedDeferred().chainDeferred(self.stopDeferred)
         defer.returnValue(smpp)
-    
+
     def handleStartError(self, error):
         self.stopDeferred.errback(error)
         return error
-        
+
     def stopService(self):
-        service.Service.stopService(self)
+        Service.stopService(self)
         if self.client.smpp:
             self.log.info("Stopping SMPP Client")
             return self.client.smpp.unbindAndDisconnect()
