+++ tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Twisted requires a package to run tests on
+"""
