+++ tests/test_smpp_protocol.py
@@ -13,22 +13,40 @@ Copyright 2009-2010 Mozes, Inc.
    See the License for the specific language governing permissions and
    limitations under the License.
 """
-import logging, binascii
-from twisted.trial import unittest
-from twisted.internet import error, reactor, defer
-from twisted.python import log
+import logging
+import binascii
+from twisted.trial.unittest import TestCase
+from twisted.internet import error, defer
 from mock import Mock, sentinel
-from smpp.pdu.error import *
-from smpp.pdu.operations import *
-from smpp.pdu.pdu_types import *
+import sys
+
+sys.path.append(".")
+sys.path.append("..")
+sys.path.append("../..")
+from smpp.pdu.error import (
+    SMPPClientConnectionCorruptedError,
+    SMPPClientSessionStateError,
+    SMPPClientError
+)
+from smpp.pdu.operations import (
+    SubmitSM,
+    GenericNack,
+    DeliverSM,
+    DeliverSMResp,
+    Unbind,
+    UnbindResp,
+    DataSM,
+    DataSMResp
+)
+from smpp.pdu.pdu_types import CommandStatus, AddrTon, AddrNpi, DeliveryFailureReason
 from smpp.twisted.config import SMPPClientConfig
 from smpp.twisted.protocol import SMPPClientProtocol, SMPPSessionStates, SMPPOutboundTxnResult, DataHandlerResponse
 
 class FakeClientError(SMPPClientError):
     pass
 
-class ProtocolTestCase(unittest.TestCase):
-    
+class ProtocolTestCase(TestCase):
+
     def getProtocolObject(self):
         smpp = SMPPClientProtocol()
         config = SMPPClientConfig(
@@ -39,10 +57,10 @@ class ProtocolTestCase(unittest.TestCase):
         )
         smpp.config = Mock(return_value=config)
         return smpp
-        
+
     def test_corruptData(self):
         smpp = self.getProtocolObject()
-        self.assertEquals('', smpp.recvBuffer)
+        self.assertEquals(b'', smpp.recvBuffer)
         smpp.sendPDU = Mock()
         smpp.cancelOutboundTransactions = Mock()
         smpp.shutdown = Mock()
@@ -57,7 +75,7 @@ class ProtocolTestCase(unittest.TestCase):
         nackResp = smpp.sendPDU.call_args[0][0]
         self.assertEquals(GenericNack(seqNum=None, status=CommandStatus.ESME_RINVCMDLEN), nackResp)
         #Causes new data received to be ignored
-        newDataHex = 'afc4'
+        newDataHex = b'afc4'
         smpp.dataReceived(binascii.a2b_hex(newDataHex))
         self.assertEquals(newDataHex, binascii.b2a_hex(smpp.recvBuffer))
         #Causes new data requests to fail immediately
@@ -70,7 +88,7 @@ class ProtocolTestCase(unittest.TestCase):
             short_message='HELLO',
         )
         return self.assertFailure(smpp.sendDataRequest(submitPdu), SMPPClientConnectionCorruptedError)
-    
+
     def test_cancelOutboundTransactions(self):
         smpp = self.getProtocolObject()
         smpp.sendPDU = Mock()
@@ -100,12 +118,12 @@ class ProtocolTestCase(unittest.TestCase):
             self.assertFailure(d1, FakeClientError),
             self.assertFailure(d2, FakeClientError),
         ])
-    
+
     def test_finish_txns(self):
         smpp = self.getProtocolObject()
         smpp.sendPDU = Mock()
         smpp.sessionState = SMPPSessionStates.BOUND_TRX
-        
+
         #setup outbound txns
         outPdu1 = SubmitSM(
             seqNum=98790,
@@ -114,7 +132,7 @@ class ProtocolTestCase(unittest.TestCase):
             short_message='HELLO1',
         )
         outRespPdu1 = outPdu1.requireAck(seqNum=outPdu1.seqNum)
-        
+
         outPdu2 = SubmitSM(
             seqNum=875,
             source_addr='mobileway',
@@ -122,34 +140,34 @@ class ProtocolTestCase(unittest.TestCase):
             short_message='HELLO1',
         )
         outRespPdu2 = outPdu2.requireAck(seqNum=outPdu2.seqNum, status=CommandStatus.ESME_RINVSRCTON)
-        
+
         outDeferred1 = smpp.startOutboundTransaction(outPdu1, 1)
         outDeferred2 = smpp.startOutboundTransaction(outPdu2, 1)
-        
+
         finishOutTxns = smpp.finishOutboundTxns()
-                
+
         #Simulate second txn having error
         smpp.endOutboundTransactionErr(outRespPdu2, FakeClientError('test'))
         #Assert txns not done yet
         self.assertFalse(finishOutTxns.called)
-        
+
         #Simulate first txn finishing
         smpp.endOutboundTransaction(outRespPdu1)
         #Assert txns are all done
         self.assertTrue(finishOutTxns.called)
-        
+
         return defer.DeferredList([
             outDeferred1,
             self.assertFailure(outDeferred2, FakeClientError),
             finishOutTxns,
         ]
         )
-    
+
     def test_graceful_unbind(self):
         smpp = self.getProtocolObject()
         smpp.sendPDU = Mock()
         smpp.sessionState = SMPPSessionStates.BOUND_TRX
-        
+
         #setup outbound txn
         outPdu = SubmitSM(
             seqNum=98790,
@@ -167,46 +185,46 @@ class ProtocolTestCase(unittest.TestCase):
             short_message='HELLO1',
         )
         inDeferred = smpp.startInboundTransaction(inPdu)
-        
+
         #Call unbind
         unbindDeferred = smpp.unbind()
-        
+
         #Assert unbind request not sent and deferred not fired
         self.assertEquals(0, smpp.sendPDU.call_count)
         self.assertFalse(unbindDeferred.called)
 
         #Simulate inbound txn finishing
         smpp.endInboundTransaction(inPdu)
-        
+
         #Assert unbind request not sent and deferred not fired
         self.assertEquals(0, smpp.sendPDU.call_count)
         self.assertFalse(unbindDeferred.called)
-        
+
         #Simulate outbound txn finishing
         smpp.endOutboundTransaction(outRespPdu)
-        
+
         #Assert unbind request was sent but deferred not yet fired
         self.assertEquals(1, smpp.sendPDU.call_count)
         sentPdu = smpp.sendPDU.call_args[0][0]
         self.assertTrue(isinstance(sentPdu, Unbind))
         self.assertFalse(unbindDeferred.called)
-        
+
         bindResp = UnbindResp(seqNum=sentPdu.seqNum)
-        
+
         #Simulate unbind_resp
         smpp.endOutboundTransaction(bindResp)
-        
+
         #Assert unbind deferred fired
         self.assertTrue(unbindDeferred.called)
         self.assertTrue(isinstance(unbindDeferred.result, SMPPOutboundTxnResult))
         expectedResult = SMPPOutboundTxnResult(smpp, sentPdu, bindResp)
         self.assertEquals(expectedResult, unbindDeferred.result)
-        
+
     def test_bind_when_not_in_open_state(self):
         smpp = self.getProtocolObject()
         smpp.sessionState = SMPPSessionStates.BOUND_TRX
         return self.assertFailure(smpp.bindAsTransmitter(), SMPPClientSessionStateError)
-        
+
     def test_unbind_when_not_bound(self):
         smpp = self.getProtocolObject()
         smpp.sessionState = SMPPSessionStates.BIND_TX_PENDING
@@ -216,13 +234,13 @@ class ProtocolTestCase(unittest.TestCase):
         smpp = self.getProtocolObject()
         smpp.sendResponse = Mock()
         smpp.disconnect = Mock()
-        
+
         smpp.sessionState = SMPPSessionStates.BOUND_TRX
         smpp.activateEnquireLinkTimer()
         self.assertNotEquals(None, smpp.enquireLinkTimer)
         smpp.onPDURequest_unbind(Unbind())
         self.assertEquals(None, smpp.enquireLinkTimer)
-        
+
     def test_sendDataRequest_when_not_bound(self):
         smpp = self.getProtocolObject()
         smpp.sessionState = SMPPSessionStates.BIND_TX_PENDING
@@ -232,7 +250,7 @@ class ProtocolTestCase(unittest.TestCase):
         smpp = self.getProtocolObject()
         smpp.sessionState = SMPPSessionStates.BOUND_TRX
         return self.assertFailure(smpp.sendDataRequest(Unbind()), SMPPClientError)
-    
+
     def test_data_handler_return_none(self):
         smpp = self.getProtocolObject()
         smpp.sendPDU = Mock()
@@ -241,7 +259,7 @@ class ProtocolTestCase(unittest.TestCase):
         self.assertEquals(1, smpp.sendPDU.call_count)
         sent = smpp.sendPDU.call_args[0][0]
         self.assertEquals(DeliverSMResp(5), sent)
-        
+
     def test_data_handler_return_status(self):
         smpp = self.getProtocolObject()
         smpp.sendPDU = Mock()
@@ -270,13 +288,3 @@ class ProtocolTestCase(unittest.TestCase):
         self.assertEquals(1, smpp.sendPDU.call_count)
         sent = smpp.sendPDU.call_args[0][0]
         self.assertEquals(DeliverSMResp(5, CommandStatus.ESME_RX_T_APPN), sent)
-
-if __name__ == '__main__':
-    observer = log.PythonLoggingObserver()
-    observer.start()
-    logging.basicConfig(level=logging.DEBUG)
-    
-    import sys
-    from twisted.scripts import trial
-    sys.argv.extend([sys.argv[0]])
-    trial.run()
